# apps/api/cmd/server/openapi.bundled.yaml
swagger: '2.0'
info:
  title: Terp API
  description: |
    RESTful API for the Terp application - Time & Resource Planning.

    ## Authentication
    All endpoints except `/health` and `/auth/*` require a valid JWT token in the Authorization header:
    ```
    Authorization: Bearer <token>
    ```

    ## Pagination
    List endpoints support cursor-based pagination:
    - `limit`: Number of items to return (default: 20, max: 100)
    - `cursor`: Cursor for next page (returned in response)

    ## Error Responses
    All errors follow RFC 7807 Problem Details format.

    ## File Structure
    This spec is organized into:
    - `paths/` - API endpoint definitions by domain
    - `schemas/` - Data models by domain
    - `responses/` - Reusable response definitions
  version: 1.0.0
  contact:
    name: API Support
  license:
    name: Proprietary
host: 'localhost:8080'
basePath: /api/v1
schemes:
  - http
  - https
consumes:
  - application/json
produces:
  - application/json
...
securityDefinitions:
  BearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: JWT token. Format "Bearer <token>"
security:
  - BearerAuth: []
paths:
...
  /employee-groups:
    get:
      tags:
        - Groups
      summary: List employee groups
      description: Returns all employee groups for the tenant.
      operationId: listEmployeeGroups
      parameters:
        - name: active_only
          in: query
          type: boolean
          description: Filter by active status
      responses:
        '200':
          description: List of employee groups
          schema:
            type: array
            items:
              $ref: '#/definitions/EmployeeGroup'
        '401':
          $ref: '#/responses/Unauthorized'
    post:
      tags:
        - Groups
      summary: Create employee group
      description: Creates a new employee group with a unique code.
      operationId: createEmployeeGroup
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateGroupRequest'
      responses:
        '201':
          description: Created employee group
          schema:
            $ref: '#/definitions/EmployeeGroup'
        '400':
          $ref: '#/responses/BadRequest'
        '401':
          $ref: '#/responses/Unauthorized'
        '409':
          description: Code already exists
          schema:
            $ref: '#/definitions/ProblemDetails'
  '/employee-groups/{id}':
    get:
      tags:
        - Groups
      summary: Get employee group by ID
      operationId: getEmployeeGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Employee group details
          schema:
            $ref: '#/definitions/EmployeeGroup'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
    patch:
      tags:
        - Groups
      summary: Update employee group
      operationId: updateEmployeeGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateGroupRequest'
      responses:
        '200':
          description: Updated employee group
          schema:
            $ref: '#/definitions/EmployeeGroup'
        '400':
          $ref: '#/responses/BadRequest'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
    delete:
      tags:
        - Groups
      summary: Delete employee group
      operationId: deleteEmployeeGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: Employee group deleted
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
  /workflow-groups:
    get:
      tags:
        - Groups
      summary: List workflow groups
      description: Returns all workflow groups for the tenant.
      operationId: listWorkflowGroups
      parameters:
        - name: active_only
          in: query
          type: boolean
          description: Filter by active status
      responses:
        '200':
          description: List of workflow groups
          schema:
            type: array
            items:
              $ref: '#/definitions/WorkflowGroup'
        '401':
          $ref: '#/responses/Unauthorized'
    post:
      tags:
        - Groups
      summary: Create workflow group
      description: Creates a new workflow group with a unique code.
      operationId: createWorkflowGroup
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateGroupRequest'
      responses:
        '201':
          description: Created workflow group
          schema:
            $ref: '#/definitions/WorkflowGroup'
        '400':
          $ref: '#/responses/BadRequest'
        '401':
          $ref: '#/responses/Unauthorized'
        '409':
          description: Code already exists
          schema:
            $ref: '#/definitions/ProblemDetails'
  '/workflow-groups/{id}':
    get:
      tags:
        - Groups
      summary: Get workflow group by ID
      operationId: getWorkflowGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Workflow group details
          schema:
            $ref: '#/definitions/WorkflowGroup'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
    patch:
      tags:
        - Groups
      summary: Update workflow group
      operationId: updateWorkflowGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateGroupRequest'
      responses:
        '200':
          description: Updated workflow group
          schema:
            $ref: '#/definitions/WorkflowGroup'
        '400':
          $ref: '#/responses/BadRequest'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
    delete:
      tags:
        - Groups
      summary: Delete workflow group
      operationId: deleteWorkflowGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: Workflow group deleted
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
  /activity-groups:
    get:
      tags:
        - Groups
      summary: List activity groups
      description: Returns all activity groups for the tenant.
      operationId: listActivityGroups
      parameters:
        - name: active_only
          in: query
          type: boolean
          description: Filter by active status
      responses:
        '200':
          description: List of activity groups
          schema:
            type: array
            items:
              $ref: '#/definitions/ActivityGroup'
        '401':
          $ref: '#/responses/Unauthorized'
    post:
      tags:
        - Groups
      summary: Create activity group
      description: Creates a new activity group with a unique code.
      operationId: createActivityGroup
      parameters:
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/CreateGroupRequest'
      responses:
        '201':
          description: Created activity group
          schema:
            $ref: '#/definitions/ActivityGroup'
        '400':
          $ref: '#/responses/BadRequest'
        '401':
          $ref: '#/responses/Unauthorized'
        '409':
          description: Code already exists
          schema:
            $ref: '#/definitions/ProblemDetails'
  '/activity-groups/{id}':
    get:
      tags:
        - Groups
      summary: Get activity group by ID
      operationId: getActivityGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '200':
          description: Activity group details
          schema:
            $ref: '#/definitions/ActivityGroup'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
    patch:
      tags:
        - Groups
      summary: Update activity group
      operationId: updateActivityGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
        - name: body
          in: body
          required: true
          schema:
            $ref: '#/definitions/UpdateGroupRequest'
      responses:
        '200':
          description: Updated activity group
          schema:
            $ref: '#/definitions/ActivityGroup'
        '400':
          $ref: '#/responses/BadRequest'
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
    delete:
      tags:
        - Groups
      summary: Delete activity group
      operationId: deleteActivityGroup
      parameters:
        - name: id
          in: path
          required: true
          type: string
          format: uuid
      responses:
        '204':
          description: Activity group deleted
        '401':
          $ref: '#/responses/Unauthorized'
        '404':
          $ref: '#/responses/NotFound'
definitions:
...
  ProblemDetails:
    type: object
    required:
      - type
      - title
      - status
    properties:
      type:
        type: string
        format: uri
        description: URI identifying the problem type
        example: 'https://api.example.com/problems/validation-error'
      title:
        type: string
        description: Short human-readable summary
        example: Validation Error
      status:
        type: integer
        description: HTTP status code
        example: 400
      detail:
        type: string
        description: Human-readable explanation
        example: The 'content' field is required
      instance:
        type: string
        format: uri
        description: URI identifying the specific occurrence
      errors:
        type: array
        description: Field-level validation errors
        items:
          type: object
          properties:
            field:
              type: string
              example: content
            message:
              type: string
              example: Field is required
...
  EmployeeGroup:
    type: object
    required:
      - id
      - tenant_id
      - code
      - name
    properties:
      id:
        type: string
        format: uuid
      tenant_id:
        type: string
        format: uuid
      code:
        type: string
        example: GRP-001
      name:
        type: string
        example: Production Workers
      description:
        type: string
        x-nullable: true
      is_active:
        type: boolean
        example: true
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  WorkflowGroup:
    type: object
    required:
      - id
      - tenant_id
      - code
      - name
    properties:
      id:
        type: string
        format: uuid
      tenant_id:
        type: string
        format: uuid
      code:
        type: string
        example: WF-001
      name:
        type: string
        example: Standard Workflow
      description:
        type: string
        x-nullable: true
      is_active:
        type: boolean
        example: true
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  ActivityGroup:
    type: object
    required:
      - id
      - tenant_id
      - code
      - name
    properties:
      id:
        type: string
        format: uuid
      tenant_id:
        type: string
        format: uuid
      code:
        type: string
        example: ACT-001
      name:
        type: string
        example: Office Activities
      description:
        type: string
        x-nullable: true
      is_active:
        type: boolean
        example: true
      created_at:
        type: string
        format: date-time
      updated_at:
        type: string
        format: date-time
  CreateGroupRequest:
    type: object
    required:
      - code
      - name
    properties:
      code:
        type: string
        minLength: 1
        maxLength: 50
      name:
        type: string
        minLength: 1
        maxLength: 255
      description:
        type: string
  UpdateGroupRequest:
    type: object
    properties:
      code:
        type: string
        minLength: 1
        maxLength: 50
      name:
        type: string
        minLength: 1
        maxLength: 255
      description:
        type: string
      is_active:
        type: boolean
...
responses:
  BadRequest:
    description: Invalid request
    schema:
      $ref: '#/definitions/ProblemDetails'
  Unauthorized:
    description: Missing or invalid authentication
    schema:
      $ref: '#/definitions/ProblemDetails'
...
  NotFound:
    description: Resource not found
    schema:
      $ref: '#/definitions/ProblemDetails'
...